feat: break apart middleware.rs
Broke `middleware.rs` into sub components that are easier to understand,
compartmentalize, and potentially share with the larger org.

There should be no impact to existing unit or run tests.

Closes #391
