+++ src/server/mod.rs
@@ -64,10 +64,10 @@ macro_rules! build_app {
             // These will wrap all outbound responses with matching status codes.
             .wrap(ErrorHandlers::new().handler(StatusCode::NOT_FOUND, ApiError::render_404))
             // These are our wrappers
-            .wrap(middleware::PreConditionCheck::new())
-            .wrap(middleware::DbTransaction::new())
-            .wrap(middleware::WeaveTimestamp::new())
-            .wrap(middleware::SentryWrapper::new())
+            .wrap(middleware::precondition::PreConditionCheck::new())
+            .wrap(middleware::db::DbTransaction::new())
+            .wrap(middleware::weave::WeaveTimestamp::new())
+            .wrap(middleware::sentry::SentryWrapper::new())
             // Followed by the "official middleware" so they run first.
             .wrap(Cors::default())
             .service(
