+++ src/web/mod.rs
@@ -11,3 +11,11 @@ pub static X_LAST_MODIFIED: &str = "x-last-modified";
 pub static X_WEAVE_TIMESTAMP: &str = "x-weave-timestamp";
 pub static X_WEAVE_NEXT_OFFSET: &str = "x-weave-next-offset";
 pub static X_WEAVE_RECORDS: &str = "x-weave-records";
+
+// Known DockerFlow commands for Ops callbacks
+pub const DOCKER_FLOW_ENDPOINTS: [&str; 4] = [
+    "/__heartbeat__",
+    "/__lbheartbeat__",
+    "/__version__",
+    "/__error__",
+];
